r2466/sofa-dev : FIX: chainHybrid with the RegularGrid. Setting of initial position done using the TransformationVisitor
WANING:  * function get of GNodes will not get objects under desactivated nodes. Like that, we can desactivate a node at initialization without problem due to the topology. Collision models under desactivated node won't be passed to the collision pipeline

	 * "Data" and "DataPtr" can be not displayed in the GUI. It is done by putting in initData, or initDataPtr as last argument a boolean. If it is false, the Data won't be displayed. It can be changed dynamically. ( setDisplayed(bool) ).

FIX: warnings

 //partial sync (15/18 changes)//

+++ modules/sofa/component/visualmodel/OglTetrahedralModel.inl
@@ -78,7 +78,7 @@ void OglTetrahedralModel<DataTypes>::drawTransparent()
         for (unsigned int i=0 ; i< 4 ; i++)
         {
             v = x[(*it)[i]];
-            glVertex3f(v[0], v[1], v[2]);
+            glVertex3f((GLfloat)v[0], (GLfloat)v[1], (GLfloat)v[2]);
         }
     }
     glEnd();
